+++ CHANGELOG.md
@@ -12,7 +12,7 @@ Bugfixes:
 ### 10.3.3 (6 November 2021)
 
 Bugfixes:
-  - LokiJS Storate must have a deterministic sort order.
+  - LokiJS Storage must have a deterministic sort order.
 ### 10.3.2 (5 November 2021)
 
 Bugfixes: